+++ lib/hooks/useApiRequest.d.ts
@@ -1,5 +1,5 @@
 export declare const useApiRequest: ({ id, options, requsetMethod, endpoint }: {
-    id?: number | undefined;
+    id?: string | number | undefined;
     options?: number | object | undefined;
     requsetMethod?: string | undefined;
     endpoint?: string | undefined;