+++ lib/hooks/useBlockRevisions.d.ts
@@ -1,7 +1,7 @@
 declare type UseCreateBlockRevision = {
     parent?: number;
-    date?: string;
-    date_gmt?: string;
+    date?: string | null;
+    date_gmt?: string | null;
     slug?: number;
     status?: 'publish' | 'future' | 'draft' | 'pending' | 'private';
     password?: string;
@@ -13,12 +13,17 @@ declare type UseRetrieveBlockRevision = {
     parent?: number;
     context?: 'view' | 'embed' | 'edit';
 };
+export declare const useRetrieveBlockRevisions: (id: number, options: UseRetrieveBlockRevision) => {
+    data: object[];
+    loading: boolean;
+    error: string | object | null;
+};
 export declare const useCreateBlockRevision: (id: number, options: UseCreateBlockRevision) => {
     data: object[];
     loading: boolean;
     error: string | object | null;
 };
-export declare const useRetrieveBlockRevision: (id: number, options: UseRetrieveBlockRevision) => {
+export declare const useRetrieveBlockRevision: (parent: number, id: number, options: UseRetrieveBlockRevision) => {
     data: object[];
     loading: boolean;
     error: string | object | null;