+++ src/hooks/useTaxonomies.ts
@@ -1,25 +1,31 @@
 import { useApiRequest } from './useApiRequest';
 
-type UseTaxonomies = {
+type UseRetrieveTaxonomies = {
   context?: 'view' | 'embed' | 'edit';
   type?: string;
 };
 
 type UseRetrieveTaxonomy = {
-  taxonomy?: string;
   context?: 'view' | 'embed' | 'edit';
 };
 
 const endpoint = 'taxonomies';
 
-export const useTaxonomies = (options: UseTaxonomies) => {
+export const useRetrieveTaxonomies = (options: UseRetrieveTaxonomies) => {
   const { data, loading, error } = useApiRequest({ options, endpoint });
 
   return { data, loading, error };
 };
 
-export const useRetrieveTaxonomy = (options: UseRetrieveTaxonomy) => {
-  const { data, loading, error } = useApiRequest({ options, endpoint });
+export const useRetrieveTaxonomy = (
+  taxonomy: string,
+  options: UseRetrieveTaxonomy
+) => {
+  const { data, loading, error } = useApiRequest({
+    id: taxonomy,
+    options,
+    endpoint
+  });
 
   return { data, loading, error };
 };