+++ src/hooks/useTypes.ts
@@ -1,24 +1,27 @@
 import { useApiRequest } from './useApiRequest';
 
-type UseTypes = {
+type UseRetrieveTypes = {
   context?: 'view' | 'embed' | 'edit';
 };
 
 type UseRetrieveType = {
-  type?: string;
   context?: 'view' | 'embed' | 'edit';
 };
 
 const endpoint = 'types';
 
-export const useTypes = (options: UseTypes) => {
+export const useRetrieveTypes = (options: UseRetrieveTypes) => {
   const { data, loading, error } = useApiRequest({ options, endpoint });
 
   return { data, loading, error };
 };
 
-export const useRetrieveType = (options: UseRetrieveType) => {
-  const { data, loading, error } = useApiRequest({ options, endpoint });
+export const useRetrieveType = (type: string, options: UseRetrieveType) => {
+  const { data, loading, error } = useApiRequest({
+    id: type,
+    options,
+    endpoint
+  });
 
   return { data, loading, error };
 };